Fix code formatting and tests
